[Feat] allow users to select crawling data type (disease, weather, environment)
